Prepare campaign for C23_WMDE_Desktop_DE_03
Ticket: https://phabricator.wikimedia.org/T342206
