+++ banners/desktop/content/BannerText.vue
@@ -7,7 +7,7 @@
 			</p>
 			<p>
 				Vielleicht kommen wir gerade ungelegen, aber dennoch: Klicken Sie jetzt bitte nicht weg! Am
-				heutigen {{ currentDayName }} bitten wir Sie bescheiden, die Unabhängigkeit von Wikipedia zu
+				heutigen {{ currentDayName }}, den {{ currentDate }}, bitten wir Sie bescheiden, die Unabhängigkeit von Wikipedia zu
 				unterstützen. <span class="wmde-banner-text-animated-highlight">Millionen Menschen nutzen Wikipedia,
 				aber 99&nbsp;% spenden nicht – sie übergehen diesen Aufruf.</span> Die meisten Menschen spenden, weil sie
 				Wikipedia nützlich finden. Die durchschnittliche Spende beträgt 22,25&nbsp;€, doch bereits 5&nbsp;€ helfen uns
@@ -19,10 +19,17 @@
 </template>
 
 <script setup lang="ts">
-import { inject } from 'vue';
+import { computed, inject } from 'vue';
 import { DynamicContent } from '@src/utils/DynamicContent/DynamicContent';
 import InfoIcon from '@src/components/Icons/InfoIcon.vue';
 
 const { currentDayName }: DynamicContent = inject( 'dynamicCampaignText' );
+
+const currentDate = computed( () => {
+	return new Intl.DateTimeFormat( 'de-DE', {
+		day: '2-digit',
+		month: 'long'
+	} ).format( new Date() );
+} );
 
 </script>
